split queue qos tests to its own file
